requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==1.1.1
langdetect==1.0.9
streamlit==1.37.0
python-dotenv==1.0.1
numpy==1.26.4
fpdf2